cleanup, doc and rename
