+++ NLog.Web/NLog.Web.csproj
@@ -45,8 +45,8 @@
   <ItemGroup>
     <Compile Include="HttpContextDiagnosticsContext.cs" />
     <Compile Include="Internal\FormatHelper.cs" />
-    <Compile Include="LayoutRenderers\HttpContextDiagnosticsContextLayoutRenderer.cs" />
     <Compile Include="LayoutRenderers\AspNetApplicationValueLayoutRenderer.cs" />
+    <Compile Include="LayoutRenderers\ApNetHttpDiagnosticsContextLayoutRenderer.cs" />
     <Compile Include="LayoutRenderers\AspNetRequestValueLayoutRenderer.cs" />
     <Compile Include="LayoutRenderers\AspNetSessionIDLayoutRenderer.cs" />
     <Compile Include="LayoutRenderers\AspNetSessionValueLayoutRenderer.cs" />
